GUI separated from Application.cpp
