+++ Application.cpp
@@ -24,6 +24,7 @@
 #include "PropertyWindow.h"
 #include <commctrl.h>
 #include "StringFunctions.h"
+#include "GuiRoot.h"
 
 #include "GUDButtonListener.h"
 #include "ModeSelectionListener.h"
@@ -53,27 +54,11 @@ PartInstance* Application::makePart()
 	return part;
 }
 
-TextButtonInstance* Application::makeTextButton()
-{
-	TextButtonInstance* part = new TextButtonInstance();
-	return part;
-}
-
-ImageButtonInstance* Application::makeImageButton(G3D::TextureRef newImage = NULL, G3D::TextureRef overImage = NULL, G3D::TextureRef downImage = NULL, G3D::TextureRef disableImage = NULL)
-{
-	ImageButtonInstance* part = new ImageButtonInstance(newImage,overImage, downImage, disableImage);
-//	instances.push_back(part);
-//	instances_2D.push_back(part);
-	return part;
-}
-
-Application::Application(HWND parentWindow) : _message(""), _messageTime(0) { //: GApp(settings,window) {
+Application::Application(HWND parentWindow) { //: GApp(settings,window) {
 	
 	std::string tempPath = ((std::string)getenv("temp")) + "/"+g_PlaceholderName;
 	CreateDirectory(tempPath.c_str(), NULL);
 	
-	_message = tempPath;
-	_messageTime = System::time();
 	_hWndMain = parentWindow;
 
 	HMODULE hThisInstance = GetModuleHandle(NULL);
@@ -167,342 +152,6 @@ void Application::deleteInstance()
 		g_usableApp->_propWindow->ClearProperties();
 }
 
-void Application::initGUI()
-{
-	TextButtonInstance* button = makeTextButton();
-	button->boxBegin = Vector2(0, -24);
-	button->boxEnd = Vector2(80, 0);
-	button->floatBottom = true;
-	button->setParent(_dataModel->getGuiRoot());
-	button->font = fntlighttrek;
-	button->textColor = Color3(0,255,255);
-	button->textOutlineColor = Color4::clear();
-	button->title = "Hopper";
-	button->fontLocationRelativeTo = Vector2(10, 3);
-	button->setAllColorsSame();
-	
-	button = makeTextButton();
-	button->boxBegin = Vector2(0, -48);
-	button->boxEnd = Vector2(80, -24);
-	button->floatBottom = true;
-	button->setParent(_dataModel->getGuiRoot());
-	button->font = fntlighttrek;
-	button->textColor = Color3(0,255,255);
-	button->textOutlineColor = Color4::clear();
-	button->title = "Controller";
-	button->fontLocationRelativeTo = Vector2(10, 3);
-	button->setAllColorsSame();
-
-	button = makeTextButton();
-	button->boxBegin = Vector2(0, -72);
-	button->boxEnd = Vector2(80, -48);
-	button->floatBottom = true;
-	button->setParent(_dataModel->getGuiRoot());
-	button->font = fntlighttrek;
-	button->textColor = Color3(0,255,255);
-	button->textOutlineColor = Color4::clear();
-	button->title = "Color";
-	button->fontLocationRelativeTo = Vector2(10, 3);
-	button->setAllColorsSame();
-
-	button = makeTextButton();
-	button->boxBegin = Vector2(0, -96);
-	button->boxEnd = Vector2(80, -72);
-	button->floatBottom = true;
-	button->setParent(_dataModel->getGuiRoot());
-	button->font = fntlighttrek;
-	button->textColor = Color3(0,255,255);
-	button->textOutlineColor = Color4::clear();
-	button->title = "Surface";
-	button->fontLocationRelativeTo = Vector2(10, 3);
-	button->setAllColorsSame();
-
-	button = makeTextButton();
-	button->boxBegin = Vector2(0, -120);
-	button->boxEnd = Vector2(80, -96);
-	button->floatBottom = true;
-	button->setParent(_dataModel->getGuiRoot());
-	button->font = fntlighttrek;
-	button->textColor = Color3(0,255,255);
-	button->boxOutlineColor = Color3(0,255,255);
-	button->title = "Model";
-	button->fontLocationRelativeTo = Vector2(10, 3);
-	button->setAllColorsSame();
-
-	button = makeTextButton();
-	button->boxBegin = Vector2(0, 0);
-	button->boxEnd = Vector2(125, 25);
-	button->setParent(_dataModel->getGuiRoot());
-	button->font = fntlighttrek;
-	button->textColor = Color3::white();
-	button->boxColor = Color4::clear();
-	button->textOutlineColor = Color4(0.5F,0.5F,0.5F,0.5F);
-	button->title = "File";
-	button->textSize = 16;
-	button->fontLocationRelativeTo = Vector2(10, 0);
-	button->setAllColorsSame();
-
-	button = makeTextButton();
-	button->boxBegin = Vector2(125, 0);
-	button->boxEnd = Vector2(250, 25);
-	button->setParent(_dataModel->getGuiRoot());
-	button->font = fntlighttrek;
-	button->textColor = Color3::white();
-	button->boxColor = Color4::clear();
-	button->textOutlineColor = Color4(0.5F,0.5F,0.5F,0.5F);
-	button->title = "Edit";
-	button->textSize = 16;
-	button->fontLocationRelativeTo = Vector2(10, 0);
-	button->setAllColorsSame();
-
-	button = makeTextButton();
-	button->boxBegin = Vector2(250, 0);
-	button->boxEnd = Vector2(375, 25);
-	button->setParent(_dataModel->getGuiRoot());
-	button->font = fntlighttrek;
-	button->textColor = Color3::white();
-	button->boxColor = Color4::clear();
-	button->textOutlineColor = Color4(0.5F,0.5F,0.5F,0.5F);
-	button->title = "View";
-	button->textSize = 16;
-	button->fontLocationRelativeTo = Vector2(10, 0);
-	button->setAllColorsSame();
-
-	button = makeTextButton();
-	button->boxBegin = Vector2(375, 0);
-	button->boxEnd = Vector2(500, 25);
-	button->setParent(_dataModel->getGuiRoot());
-	button->font = fntlighttrek;
-	button->textColor = Color3::white();
-	button->boxColor = Color4::clear();
-	button->textOutlineColor = Color4(0.5F,0.5F,0.5F,0.5F);
-	button->title = "Insert";
-	button->textSize = 16;
-	button->fontLocationRelativeTo = Vector2(10, 0);
-	button->setAllColorsSame();
-
-	button = makeTextButton();
-	button->boxBegin = Vector2(500, 0);
-	button->boxEnd = Vector2(625, 25);
-	button->setParent(_dataModel->getGuiRoot());
-	button->font = fntlighttrek;
-	button->textColor = Color3::white();
-	button->boxColor = Color4::clear();
-	button->textOutlineColor = Color4(0.5F,0.5F,0.5F,0.5F);
-	button->title = "Format";
-	button->textSize = 16;
-	button->fontLocationRelativeTo = Vector2(10, 0);
-	button->setAllColorsSame();
-
-
-
-	button = makeTextButton();
-	button->boxBegin = Vector2(0,215);
-	button->boxEnd = Vector2(80,235);
-	button->textOutlineColor = Color4(0.5F,0.5F,0.5F,0.5F);
-	button->textColor = Color3::white();
-	button->boxColor = Color4::clear();
-	button->textSize = 12;
-	button->title = "Group";
-	button->name = "Group";
-	button->setAllColorsSame();	
-	button->textColorDis = Color3(0.8F,0.8F,0.8F);
-	button->font = fntlighttrek;
-	button->fontLocationRelativeTo = Vector2(10, 0);
-	button->setParent(_dataModel->getGuiRoot());
-
-
-	button = makeTextButton();
-	button->boxBegin = Vector2(0,240);
-	button->boxEnd = Vector2(80,260);
-	button->textOutlineColor = Color4(0.5F,0.5F,0.5F,0.5F);
-	button->textColor = Color3::white();
-	button->boxColor = Color4::clear();
-	button->textSize = 12;
-	button->title = "UnGroup";
-	button->name = "UnGroup";
-	button->setAllColorsSame();
-	button->textColorDis = Color3(0.8F,0.8F,0.8F);
-	button->font = fntlighttrek;
-	button->fontLocationRelativeTo = Vector2(10, 0);
-	button->setParent(_dataModel->getGuiRoot());
-
-	button = makeTextButton();
-	button->boxBegin = Vector2(0,265);
-	button->boxEnd = Vector2(80,285);
-	button->textOutlineColor = Color4(0.5F,0.5F,0.5F,0.5F);
-	button->textColor = Color3::white();
-	button->boxColor = Color4::clear();
-	button->textSize = 12;
-	button->title = "Duplicate";
-	button->setAllColorsSame();
-	button->textColorDis = Color3(0.8F,0.8F,0.8F);
-	button->font = fntlighttrek;
-	button->fontLocationRelativeTo = Vector2(10, 0);
-	button->setParent(_dataModel->getGuiRoot());
-	button->name = "Duplicate";
-	button->setButtonListener(new GUDButtonListener());
-
-	ImageButtonInstance* instance = makeImageButton(
-		Texture::fromFile(GetFileInPath("/content/images/Run.png")),
-		Texture::fromFile(GetFileInPath("/content/images/Run_ovr.png")),
-		Texture::fromFile(GetFileInPath("/content/images/Run_dn.png")));
-	instance->name = "go";
-	instance->size = Vector2(65,65);
-	instance->position = Vector2(6.5, 25);
-	instance->setParent(_dataModel->getGuiRoot());
-
-
-	instance = makeImageButton(
-		Texture::fromFile(GetFileInPath("/content/images/ArrowTool.png")),
-		Texture::fromFile(GetFileInPath("/content/images/ArrowTool_ovr.png")),
-		Texture::fromFile(GetFileInPath("/content/images/ArrowTool_dn.png")),
-		Texture::fromFile(GetFileInPath("/content/images/ArrowTool_ds.png")));
-	instance->size = Vector2(50,50);
-	instance->position = Vector2(15, 90);
-	instance->setParent(_dataModel->getGuiRoot());
-	instance->name = "Cursor";
-	instance->setButtonListener(new ModeSelectionListener());
-
-	instance = makeImageButton(Texture::fromFile(GetFileInPath("/content/images/ScaleTool.png")),Texture::fromFile(GetFileInPath("/content/images/ScaleTool_ovr.png")),Texture::fromFile(GetFileInPath("/content/images/ScaleTool_dn.png")),Texture::fromFile(GetFileInPath("/content/images/ScaleTool_ds.png")));
-	instance->size = Vector2(40,40);
-	instance->position = Vector2(0, 140);
-	instance->setParent(_dataModel->getGuiRoot());
-	instance->name = "Resize";
-	instance->setButtonListener(new ModeSelectionListener());
-	
-
-	instance = makeImageButton(
-		Texture::fromFile(GetFileInPath("/content/images/MoveTool.png")),
-		Texture::fromFile(GetFileInPath("/content/images/MoveTool_ovr.png")),
-		Texture::fromFile(GetFileInPath("/content/images/MoveTool_dn.png")),
-		Texture::fromFile(GetFileInPath("/content/images/MoveTool_ds.png")));
-	instance->size = Vector2(40,40);
-	instance->position = Vector2(40, 140);
-	instance->setParent(_dataModel->getGuiRoot());
-	instance->name = "Arrows";
-	instance->setButtonListener(new ModeSelectionListener());
-
-	instance = makeImageButton(
-		Texture::fromFile(GetFileInPath("/content/images/SelectionRotate.png")),
-		Texture::fromFile(GetFileInPath("/content/images/SelectionRotate_ovr.png")),
-		NULL,
-		Texture::fromFile(GetFileInPath("/content/images/SelectionRotate_ds.png")));
-	instance->size = Vector2(30,30);
-	instance->position = Vector2(10, 175);
-	instance->setParent(_dataModel->getGuiRoot());
-	instance->name = "Rotate";
-	instance->setButtonListener(new RotateButtonListener());
-
-	instance = makeImageButton(
-		Texture::fromFile(GetFileInPath("/content/images/SelectionTilt.png")),
-		Texture::fromFile(GetFileInPath("/content/images/SelectionTilt_ovr.png")),
-		NULL,
-		Texture::fromFile(GetFileInPath("/content/images/SelectionTilt_ds.png")));
-	instance->size = Vector2(30,30);
-	instance->position = Vector2(40, 175);
-	instance->setParent(_dataModel->getGuiRoot());
-	instance->name = "Tilt";
-	instance->setButtonListener(new RotateButtonListener());
-
-
-	instance = makeImageButton(
-		Texture::fromFile(GetFileInPath("/content/images/Delete.png")),
-		Texture::fromFile(GetFileInPath("/content/images/Delete_ovr.png")),
-		Texture::fromFile(GetFileInPath("/content/images/Delete_dn.png")),
-		Texture::fromFile(GetFileInPath("/content/images/Delete_ds.png")));
-	instance->size = Vector2(40,46);
-	instance->position = Vector2(20, 284);
-	instance->setParent(_dataModel->getGuiRoot());
-	instance->name = "Delete";
-	instance->setButtonListener(new DeleteListener());
-
-	instance = makeImageButton(
-		Texture::fromFile(GetFileInPath("/content/images/CameraZoomIn.png")),
-		Texture::fromFile(GetFileInPath("/content/images/CameraZoomIn_ovr.png")),
-		Texture::fromFile(GetFileInPath("/content/images/CameraZoomIn_dn.png")));
-	instance->size = Vector2(34,25);
-	instance->floatBottom = true;
-	instance->floatRight = true;
-	instance->position = Vector2(-77, -90);
-	instance->setParent(_dataModel->getGuiRoot());
-	instance->name = "ZoomIn";
-	instance->setButtonListener(new CameraButtonListener());
-
-	instance = makeImageButton(
-		Texture::fromFile(GetFileInPath("/content/images/CameraZoomOut.png")),
-		Texture::fromFile(GetFileInPath("/content/images/CameraZoomOut_ovr.png")),
-		Texture::fromFile(GetFileInPath("/content/images/CameraZoomOut_dn.png")));
-	instance->size = Vector2(34,26);
-	instance->floatBottom = true;
-	instance->floatRight = true;
-	instance->position = Vector2(-77, -31);
-	instance->setParent(_dataModel->getGuiRoot());
-	instance->name = "ZoomOut";
-	instance->setButtonListener(new CameraButtonListener());
-
-	instance = makeImageButton(
-		Texture::fromFile(GetFileInPath("/content/images/CameraPanLeft.png")),
-		Texture::fromFile(GetFileInPath("/content/images/CameraPanLeft_ovr.png")),
-		Texture::fromFile(GetFileInPath("/content/images/CameraPanLeft_dn.png")));
-	instance->size = Vector2(34,34);
-	instance->floatBottom = true;
-	instance->floatRight = true;
-	instance->position = Vector2(-110, -50);
-	instance->setParent(_dataModel->getGuiRoot());
-	instance->name = "PanLeft";
-	instance->setButtonListener(new CameraButtonListener());
-
-	instance = makeImageButton(
-		Texture::fromFile(GetFileInPath("/content/images/CameraPanRight.png")),
-		Texture::fromFile(GetFileInPath("/content/images/CameraPanRight_ovr.png")),
-		Texture::fromFile(GetFileInPath("/content/images/CameraPanRight_dn.png")));
-	instance->size = Vector2(34,34);
-	instance->floatBottom = true;
-	instance->floatRight = true;
-	instance->position = Vector2(-45, -50);
-	instance->setParent(_dataModel->getGuiRoot());
-	instance->name = "PanRight";
-	instance->setButtonListener(new CameraButtonListener());
-
-	instance = makeImageButton(
-		Texture::fromFile(GetFileInPath("/content/images/CameraCenter.png")),
-		Texture::fromFile(GetFileInPath("/content/images/CameraCenter_ovr.png")),
-		Texture::fromFile(GetFileInPath("/content/images/CameraCenter_dn.png")));
-	instance->size = Vector2(34,20);
-	instance->floatBottom = true;
-	instance->floatRight = true;
-	instance->position = Vector2(-77, -60);
-	instance->setParent(_dataModel->getGuiRoot());
-	instance->name = "CenterCam";
-	instance->setButtonListener(new CameraButtonListener());
-
-	instance = makeImageButton(
-		Texture::fromFile(GetFileInPath("/content/images/CameraTiltUp.png")),
-		Texture::fromFile(GetFileInPath("/content/images/CameraTiltUp_ovr.png")),
-		Texture::fromFile(GetFileInPath("/content/images/CameraTiltUp_dn.png")));
-	instance->size = Vector2(24,24);
-	instance->floatBottom = true;
-	instance->floatRight = true;
-	instance->position = Vector2(-105, -75);
-	instance->setParent(_dataModel->getGuiRoot());
-	instance->name = "TiltUp";
-	instance->setButtonListener(new CameraButtonListener());
-
-	instance = makeImageButton(
-		Texture::fromFile(GetFileInPath("/content/images/CameraTiltDown.png")),
-		Texture::fromFile(GetFileInPath("/content/images/CameraTiltDown_ovr.png")),
-		Texture::fromFile(GetFileInPath("/content/images/CameraTiltDown_dn.png")));
-	instance->size = Vector2(24,24);
-	instance->floatBottom = true;
-	instance->floatRight = true;
-	instance->position = Vector2(-40, -75);
-	instance->setParent(_dataModel->getGuiRoot());
-	instance->name = "TiltDown";
-	instance->setButtonListener(new CameraButtonListener());
-}
-
-
 void Application::onInit()  {
 	
     // Called before Application::run() beings
@@ -510,10 +159,10 @@ void Application::onInit()  {
 	_dataModel = new DataModelInstance();
 	_dataModel->setParent(NULL);
 	_dataModel->name = "undefined";
-	_dataModel->font = fntdominant;
+	_dataModel->font = g_fntdominant;
 	g_dataModel = _dataModel;
 	
-	initGUI();
+	//initGUI();
 
 #ifdef LEGACY_LOAD_G3DFUN_LEVEL
 	PartInstance* test = makePart();
@@ -648,39 +297,7 @@ void Application::onSimulation(RealTime rdt, SimTime sdt, SimTime idt) {
 
 	
 		
-		Instance * obj6 = _dataModel->getGuiRoot()->findFirstChild("Delete");
-		Instance * obj = _dataModel->getGuiRoot()->findFirstChild("Duplicate");
-		Instance * obj2 = _dataModel->getGuiRoot()->findFirstChild("Group");
-		Instance * obj3 = _dataModel->getGuiRoot()->findFirstChild("UnGroup");
-		Instance * obj4 = _dataModel->getGuiRoot()->findFirstChild("Rotate");
-		Instance * obj5 = _dataModel->getGuiRoot()->findFirstChild("Tilt");
-		if(obj != NULL && obj2 != NULL && obj3 != NULL && obj4 !=NULL && obj5 != NULL && obj6 != NULL)
-		{
-			BaseButtonInstance* button = (BaseButtonInstance*)obj;
-			BaseButtonInstance* button2 = (BaseButtonInstance*)obj2;
-			BaseButtonInstance* button3 = (BaseButtonInstance*)obj3;
-			BaseButtonInstance* button4 = (BaseButtonInstance*)obj4;
-			BaseButtonInstance* button5 = (BaseButtonInstance*)obj5;
-			BaseButtonInstance* button6 = (BaseButtonInstance*)obj6;
-			button->disabled = true;
-			button2->disabled = true;
-			button3->disabled = true;
-			button4->disabled = true;
-			button5->disabled = true;
-			button6->disabled = true;
-			for(size_t i = 0; i < g_selectedInstances.size(); i++)
-				if(g_selectedInstances.at(i)->canDelete)
-				{
-					button->disabled = false;
-					button2->disabled = false;
-					button3->disabled = false;
-					button4->disabled = false;
-					button5->disabled = false;
-					button6->disabled = false;
-					break;
-				}
-		}
-
+	_dataModel->getGuiRoot()->update();
 
 	if(_dataModel->name != _title)
 	{
@@ -730,9 +347,7 @@ void Application::onUserInput(UserInput* ui) {
 	*/
 	if(GetHoldKeyState(VK_F8))
 	{
-		_messageTime = System::time();
-		_message = "FOV Set to 10";
-		
+		_dataModel->getGuiRoot()->setDebugMessage("FOV Set to 10", System::time());	
 	}
 	//}
 
@@ -810,19 +425,15 @@ void Application::makeFlag(Vector3 &vec, RenderDevice* &rd)
 	//I know how i will approach this now
 }
 
+void Application::setMode(int mode)
+{
+	_mode = mode;
+}
 int Application::getMode()
 {
 	return _mode;
 }
 
-void Application::drawButtons(RenderDevice* rd)
-{
-	rd->pushState();
-	rd->beforePrimitive();
-			_dataModel->getGuiRoot()->render(rd);
-	rd->afterPrimitive();
-	rd->popState();
-}
 
 void Application::drawOutline(Vector3 from, Vector3 to, RenderDevice* rd, LightingParameters lighting, Vector3 size, Vector3 pos, CoordinateFrame c)
 {
@@ -986,7 +597,7 @@ void Application::onGraphics(RenderDevice* rd) {
 			{
 			Vector3 size = part->getSize();
 			Vector3 pos = part->getPosition();
-			drawOutline(Vector3(0+size.x/2, 0+size.y/2, 0+size.z/2) ,Vector3(0-size.x/2,0-size.y/2,0-size.z/2), rd, lighting, Vector3(size.x/2, size.y/2, size.z/2), Vector3(pos.x/2, pos.y/2, pos.z/2), part->getCFrameRenderBased());
+			drawOutline(Vector3(0+size.x/2, 0+size.y/2, 0+size.z/2) ,Vector3(0-size.x/2,0-size.y/2,0-size.z/2), rd, lighting, Vector3(size.x/2, size.y/2, size.z/2), Vector3(pos.x, pos.y, pos.z), part->getCFrameRenderBased());
 			}
 		}
 	}
@@ -1009,33 +620,7 @@ void Application::onGraphics(RenderDevice* rd) {
         sky->renderLensFlare(renderDevice, lighting);
     }
 	renderDevice->push2D();
-		if(System::time() - 3 < _messageTime)
-		{
-			fntdominant->draw2D(rd, _message, Vector2((rd->getWidth()/2)-(fntdominant->get2DStringBounds(_message, 20).x/2),(rd->getHeight()/2)-(fntdominant->get2DStringBounds(_message, 20).y/2)), 20, Color3::yellow(), Color3::black());
-		}
-
-		//TODO--Move these to their own instance
-
-		std::stringstream stream;
-		stream << std::fixed << std::setprecision(1) << _dataModel->getLevel()->timer;
-		fntdominant->draw2D(rd, "Timer: " + stream.str(), Vector2(rd->getWidth() - 120, 25), 20, Color3::fromARGB(0x81C518), Color3::black());
-		fntdominant->draw2D(rd, "Score: " + Convert(_dataModel->getLevel()->score), Vector2(rd->getWidth() - 120, 50), 20, Color3::fromARGB(0x81C518), Color3::black());
-		
-		//GUI Boxes	
-		Draw::box(G3D::Box(Vector3(0,25,0),Vector3(80,355,0)),rd,Color4(0.6F,0.6F,0.6F,0.4F), Color4(0,0,0,0));
-		Draw::box(G3D::Box(Vector3(rd->getWidth() - 120,rd->getHeight() - 120,0),Vector3(rd->getWidth(),rd->getHeight(),0)),rd,Color4(0.6F,0.6F,0.6F,0.4F), Color4(0,0,0,0));
-		
-		//Camera menu title
-		fntlighttrek->draw2D(rd, "CameraMenu", Vector2(rd->getWidth()-(fntlighttrek->get2DStringBounds("CameraMenu", 14).x+1),rd->getHeight() - 120), 14, Color3::white(), Color4(0.5F,0.5F,0.5F,0.5F));
-
-		//Tools menu
-		Draw::box(G3D::Box(Vector3(5, 210,0),Vector3(75, 210,0)),rd,Color4(0.6F,0.6F,0.6F,0.4F), Color4(0.6F,0.6F,0.6F,0.4F));
-		fntlighttrek->draw2D(rd,"MENU", Vector2(10,332), 14, Color3::white(), Color4(0.5F,0.5F,0.5F,0.5F));
-
-		
-		drawButtons(rd);
-
-		_dataModel->drawMessage(rd);
+		_dataModel->getGuiRoot()->renderGUI(renderDevice);
 		rd->pushState();
 			rd->beforePrimitive();
 
@@ -1110,19 +695,9 @@ void Application::onMouseLeftPressed(HWND hwnd,int x,int y)
 
 	std::cout << "Click: " << x << "," << y << std::endl;
 
-	bool onGUI = false;
-	std::vector<Instance*> instances_2D = _dataModel->getGuiRoot()->getAllChildren();
-	for(size_t i = 0; i < instances_2D.size(); i++)
-	{
-		if(BaseButtonInstance* button = dynamic_cast<BaseButtonInstance*>(instances_2D.at(i)))
-		{
-			if(button->mouseInButton(x,y, renderDevice))
-			{
-				onGUI = true;
-				break;
-			}
-		}
-	}
+	bool onGUI = _dataModel->getGuiRoot()->mouseInGUI(renderDevice, x, y);
+	
+	
 	if(!onGUI)
 	{
 		testRay = cameraController.getCamera()->worldRay(_dataModel->mousex, _dataModel->mousey, renderDevice->getViewport());
@@ -1159,7 +734,7 @@ void Application::onMouseLeftPressed(HWND hwnd,int x,int y)
 							g_selectedInstances.clear();
 							g_selectedInstances.push_back(test);
 						}
-						_propWindow->UpdateSelected(test);
+						selectInstance(test, _propWindow);
 						//_message = "Dragging = true.";
 						//_messageTime = System::time();
 						//_dragging = true;
@@ -1168,34 +743,32 @@ void Application::onMouseLeftPressed(HWND hwnd,int x,int y)
 			}		
 		}
 		if(!objFound)
-		{
-			g_selectedInstances.clear();
-			g_selectedInstances.push_back(_dataModel);
-			_propWindow->UpdateSelected(_dataModel);
-			
-		}
+			selectInstance(_dataModel,_propWindow);
 	}
 }
-void Application::onMouseLeftUp(int x,int y)
+
+void Application::selectInstance(Instance* selectedInstance, PropertyWindow* propWindow)
+{
+	g_selectedInstances.clear();
+	g_selectedInstances.push_back(selectedInstance);
+	propWindow->UpdateSelected(selectedInstance);
+
+}
+G3D::RenderDevice* Application::getRenderDevice()
+{
+	return renderDevice;
+}
+
+void Application::onMouseLeftUp(G3D::RenderDevice* renderDevice, int x, int y)
 {
 	//std::cout << "Release: " << x << "," << y << std::endl;
-
+	_dataModel->getGuiRoot()->onMouseLeftUp(renderDevice, x, y);
 	_dragging = false;
 	//_message = "Dragging = false.";
 	//_messageTime = System::time();
-	std::vector<Instance*> instances_2D = _dataModel->getGuiRoot()->getAllChildren();
-	//std::vector<Instance*> instances = _dataModel->getWorkspace()->getAllChildren();
-	for(size_t i = 0; i < instances_2D.size(); i++)
-	{
-		if(BaseButtonInstance* button = dynamic_cast<BaseButtonInstance*>(instances_2D[i]))
-		{
-			if(button->mouseInButton(x, y, renderDevice))
-			{
-				button->onMouseClick();
-			}
-		}
-	}
+
 }
+
 void Application::onMouseRightPressed(int x,int y)
 {
 }
@@ -1243,8 +816,6 @@ void Application::run() {
 	cursorOvr = Texture::fromFile(GetFileInPath("/content/images/DragCursor.png"));
 	Globals::surface = Texture::fromFile(GetFileInPath("/content/images/surfacebr.png"));
 	Globals::surfaceId = Globals::surface->getOpenGLID();
-	fntdominant = GFont::fromFile(GetFileInPath("/content/font/dominant.fnt"));
-	fntlighttrek = GFont::fromFile(GetFileInPath("/content/font/lighttrek.fnt"));
 	cameraSound = GetFileInPath("/content/sounds/SWITCH3.wav");
 	clickSound = GetFileInPath("/content/sounds/switch.wav");
 	dingSound = GetFileInPath("/content/sounds/electronicpingshort.wav");
