+++ ErrorFunctions.h
@@ -1,3 +1,4 @@
+#pragma once
 #include <string>
 
 void OnError(int err, std::string msg = "");
