+++ G3DTest.sln
@@ -1,7 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 9.00
-# Visual Studio 2005
+# Visual C++ Express 2005
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "G3DTest", "G3DTest.vcproj", "{6C4D6EEF-B1D1-456A-B850-92CAB17124BE}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "graphics3D", "..\..\..\..\SDK\VisualC\VS2005\g3d-src-6_10\source\graphics3D.vcproj", "{FD379372-8DB2-4C2C-A4D8-4694BAA94640}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GLG3D", "..\..\..\..\SDK\VisualC\VS2005\g3d-src-6_10\source\GLG3D.vcproj", "{EF60D11E-36A7-43EB-B198-DD3906DCFDBF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,6 +17,14 @@ Global
 		{6C4D6EEF-B1D1-456A-B850-92CAB17124BE}.Debug|Win32.Build.0 = Debug|Win32
 		{6C4D6EEF-B1D1-456A-B850-92CAB17124BE}.Release|Win32.ActiveCfg = Release|Win32
 		{6C4D6EEF-B1D1-456A-B850-92CAB17124BE}.Release|Win32.Build.0 = Release|Win32
+		{FD379372-8DB2-4C2C-A4D8-4694BAA94640}.Debug|Win32.ActiveCfg = Debug|Win32
+		{FD379372-8DB2-4C2C-A4D8-4694BAA94640}.Debug|Win32.Build.0 = Debug|Win32
+		{FD379372-8DB2-4C2C-A4D8-4694BAA94640}.Release|Win32.ActiveCfg = Release|Win32
+		{FD379372-8DB2-4C2C-A4D8-4694BAA94640}.Release|Win32.Build.0 = Release|Win32
+		{EF60D11E-36A7-43EB-B198-DD3906DCFDBF}.Debug|Win32.ActiveCfg = Debug|Win32
+		{EF60D11E-36A7-43EB-B198-DD3906DCFDBF}.Debug|Win32.Build.0 = Debug|Win32
+		{EF60D11E-36A7-43EB-B198-DD3906DCFDBF}.Release|Win32.ActiveCfg = Release|Win32
+		{EF60D11E-36A7-43EB-B198-DD3906DCFDBF}.Release|Win32.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
