+++ Instance.h
@@ -11,6 +11,7 @@ public:
 	virtual ~Instance(void);
 	std::string name;
 	virtual void render(RenderDevice*);
+	virtual void update();
 	std::vector<Instance*> children; // All children.
 	std::string getClassName();
 	Instance* findFirstChild(std::string);
