+++ PropertyWindow.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "Instance.h"
+
 class PropertyWindow {
 public:
 	PropertyWindow(int x, int y, int sx, int sy, HMODULE hThisInstance);
