+++ RotateButtonListener.cpp
@@ -1,3 +1,4 @@
+#include "BaseButtonInstance.h"
 #include "RotateButtonListener.h"
 #include "Globals.h"
 #include "AudioPlayer.h"
